Add name qualifier

+++ src/main/java/com/workbeattalent/sport_club/cycling/CreateCyclingTour.java
@@ -20,7 +20,7 @@ public record CreateCyclingTour(
         LocalDateTime completedAt,
 
         @Positive
-        Double distance,
+        Float distance,
 
         Location location
 ) {
